+++ config/webpack.config.dev.js
@@ -3,20 +3,20 @@ import autoprefixer from 'autoprefixer';
 import webpack from 'webpack';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 
-import { createEmbedEntryPoints, createEmbedPlugins } from './embedBuilders';
+// import { createEmbedEntryPoints, createEmbedPlugins } from './embedBuilders';
 import paths from './paths';
 import config from '../src/config/default';
 
 module.exports = {
   devtool: 'eval',
   entry: {
-    ...createEmbedEntryPoints(),
+    // ...createEmbedEntryPoints(),
     app: [
       require.resolve('webpack-dev-server/client'),
       require.resolve('webpack/hot/dev-server'),
       path.join(paths.appSrc, 'index'),
     ],
-    embed: [
+    gw2aEmbed: [
       require.resolve('webpack-dev-server/client'),
       require.resolve('webpack/hot/dev-server'),
       path.join(paths.embedSrc, 'index'),
@@ -27,6 +27,7 @@ module.exports = {
     path: paths.appBuild,
     pathinfo: true,
     filename: '[name].js',
+    chunkFilename: '[name]-chunk.js',
     publicPath: '/',
   },
   resolve: {
@@ -73,7 +74,7 @@ module.exports = {
     return [autoprefixer];
   },
   plugins: [
-    ...createEmbedPlugins(),
+    // ...createEmbedPlugins(),
     new HtmlWebpackPlugin({
       ...config,
       inject: true,
@@ -84,8 +85,8 @@ module.exports = {
       ...config,
       inject: true,
       template: paths.embedsHtml,
-      filename: 'embeds/index.html',
-      chunks: ['embed'],
+      filename: 'embeds/example/index.html',
+      chunks: ['gw2aEmbed'],
     }),
     new webpack.DefinePlugin({
       'process.env.NODE_ENV': '"development"',