+++ src/Base.js
@@ -4,6 +4,7 @@ import 'airbnb-js-shims';
 import 'normalize.css';
 import 'lib/i18n';
 import 'assets/fonts/menomonia.css';
+import 'assets/fonts/opensans.css';
 
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';